requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
kafka-python==2.0.2
pip==24.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
pip==24.2
py4j==0.10.9.7
pyspark==3.5.3
setuptools==75.1.0
wheel==0.44.0